# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sporteam API",
    "version": "2.0.0",
    "description": "API RESTful para la gestión de jugadores, entrenadores y clubs de fútbol amateur",
    "contact": {
      "name": "Roman Dono Perez"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "Token JWT de autenticación"
      }
    },
    "schemas": {
...
      "Jugador": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "estadoDeportivo": {
            "type": "string",
            "nullable": true
          },
          "clubId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "jugador": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "nombreDeportivo": {
                "type": "string",
                "nullable": true
              },
              "fechaNacimiento": {
                "type": "string",
                "format": "date",
                "nullable": true
              },
              "lateralidad": {
                "type": "string",
                "nullable": true
              },
              "demarcacion": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "altura": {
                "type": "number",
                "nullable": true
              },
              "peso": {
                "type": "number",
                "nullable": true
              }
            }
          },
          "estadisticas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Estadistica"
            }
          }
        }
      },
      "JugadorInput": {
        "type": "object",
        "required": [
          "nombre",
          "apellidos",
          "email",
          "password"
        ],
        "properties": {
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "JUGADOR_ROLE"
            ]
          },
          "estadoDeportivo": {
            "type": "string"
          },
          "zona": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "club": {
            "type": "string"
          },
          "nombreDeportivo": {
            "type": "string"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date"
          },
          "lateralidad": {
            "type": "string",
            "enum": [
              "zurdo",
              "diestro",
              "ambidiestro"
            ]
          },
          "demarcacion": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "altura": {
            "type": "number"
          },
          "peso": {
            "type": "number"
          }
        }
      },
...
      "Estadistica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "temporadaId": {
            "type": "string",
            "format": "uuid"
          },
          "partidosJugados": {
            "type": "integer"
          },
          "goles": {
            "type": "integer"
          },
          "asistencias": {
            "type": "integer"
          },
          "tarjetasAmarillas": {
            "type": "integer"
          },
          "tarjetasRojas": {
            "type": "integer"
          },
          "temporada": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "anho": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/jugadores": {
      "get": {
        "tags": [
          "Jugadores"
        ],
        "summary": "Listar jugadores con paginación",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de jugadores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "jugadores": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Jugador"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al recuperar jugadores"
          }
        }
      }
    },
    "/api/jugadores/{termino}": {
      "get": {
        "tags": [
          "Jugadores"
        ],
        "summary": "Buscar jugadores por término",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "termino",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Término de búsqueda (nombre, apellidos, email)"
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de búsqueda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "resultados": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Jugador"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jugador/{id}": {
      "get": {
        "tags": [
          "Jugadores"
        ],
        "summary": "Obtener jugador por ID de usuario",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jugador encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "jugador": {
                      "$ref": "#/components/schemas/Jugador"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Jugador no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Jugadores"
        ],
        "summary": "Actualizar un jugador",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombreDeportivo": {
                    "type": "string"
                  },
                  "fechaNacimiento": {
                    "type": "string",
                    "format": "date"
                  },
                  "lateralidad": {
                    "type": "string"
                  },
                  "demarcacion": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "altura": {
                    "type": "number"
                  },
                  "peso": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jugador actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "jugador": {
                      "$ref": "#/components/schemas/Jugador"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Jugador no encontrado"
          }
        }
      }
    },
    "/api/jugador": {
      "post": {
        "tags": [
          "Jugadores"
        ],
        "summary": "Crear un nuevo jugador (usuario + perfil)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JugadorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jugador creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "jugador": {
                      "$ref": "#/components/schemas/Jugador"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          }
        }
      }
    },
    "/api/jugadores/zona/{idZona}": {
      "get": {
        "tags": [
          "Jugadores"
        ],
        "summary": "Listar jugadores por zona",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idZona",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jugadores filtrados por zona",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "jugadores": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Jugador"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}